requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.1.1
email-validator==2.2.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
requests==2.32.4
jinja2==3.1.6
starlette
asyncpg
pydantic==2.8.2
passlib[bcrypt]==1.7.4






